move test to openid
this is easier to handle than doing it in jwt, because of how the
dependencies work
